Fix UniqueKeyEntity setter

+++ src/main/java/io/split/android/client/storage/db/impressions/unique/UniqueKeyEntity.java
@@ -78,7 +78,7 @@ public class UniqueKeyEntity implements Identifiable {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
